myhome_1/forms.py
from django import forms
import datetime
from tempus_dominus.widgets import DatePicker, TimePicker, DateTimePicker
from .models import gazoline, orbi_tmp, electricity


class DateForm(forms.Form):
    date = forms.DateTimeField(
        input_formats=['%d/%m/%Y %H:%M'],
        widget=forms.DateTimeInput(attrs={
            'class': 'form-control datetimepicker-input',
            'data-target': '#datetimepicker1'
        })
    )


class MyForm(forms.Form):
    date_field = forms.DateField(widget=DatePicker())
    date_field_required_with_min_max_date = forms.DateField(
        required=True,
        widget=DatePicker(
            options={
                'minDate': '2009-01-20',
                'maxDate': '2017-01-20',
            },
        ),
    )
    time_field = forms.TimeField(
        widget=TimePicker(
            options={
               # 'enabledHours': [9, 10, 11, 12, 13, 14, 15, 16],
            },
            attrs={
                'input_toggle': True,
                'input_group': True,
            },
        ),
    )
    datetime_field = forms.DateTimeField(
        widget=DateTimePicker(
            options={
                'minDate': (
                    datetime.date.today() + datetime.timedelta(days=1)
                ).strftime(
                    '%Y-%m-%d'
                ),  # Tomorrow
                'useCurrent': True,
                'collapse': False,
            },
            attrs={
                'append': 'fa fa-calendar',
                'icon_toggle': True,
            }
        ),
    )


class gazForm(forms.ModelForm):
    class Meta:
        model = gazoline
        fields = '__all__'


class orbiForm(forms.ModelForm):
    class Meta:
        model = orbi_tmp
        fields = '__all__'
        
        
        
class ElectricityForm(forms.ModelForm):
    class Meta:
        model = electricity
        fields = '__all__'