requirements.txt
amqp==2.6.1
anyjson==0.3.3
billiard==3.6.3.0
celery==4.4.7
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
Django==2.2.19
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-chart-tools==1.0
django-chartjs==2.0.0
django-crispy-forms==1.8.1
django-debug-toolbar==0.8.5
django-extensions==2.2.8
django-filter==2.1.0
django-filters==0.2.1
django-google-charts==0.1.1
django-json==0.27
django-qsstats-magic==1.0.0
django-tempus-dominus==5.1.2.4
django-timezone-field==4.1.2
djangorestframework==3.9.2
djangorestframework-filter==0.2.2
idna==2.9
importlib-metadata==2.1.1
kiwisolver==1.1.0
kombu==4.6.11
Markdown==3.0.1
matplotlib==3.0.3
multitasking==0.0.9
numpy==1.18.2
paho-mqtt==1.4.0
pandas==0.25.3
Pillow==7.0.0
psutil==5.6.7
psycopg2==2.8.6
pyparsing==2.4.6
pysqlite3==0.2.1
python-crontab==2.5.1
python-dateutil==2.8.0
pytz==2018.9
PyYAML==3.13
redis==3.5.3
requests==2.23.0
six==1.12.0
sqlparse==0.4.1
urllib3==1.25.8
vine==1.3.0
yfinance==0.1.54
zipp==1.2.0
